🚚 Rename mockSession to mockUsers
